add sidenav to profile
