+++ apps/web/src/app/(admin)/admin/layout.tsx
@@ -4,9 +4,9 @@ import { type Metadata } from "next/types";
 import { getAuth } from "@echo-webkom/auth";
 
 import { Footer } from "@/components/footer";
+import { SidebarLayout } from "@/components/sidebar-layout";
 import { SiteHeader } from "@/components/site-header";
 import { isWebkom } from "@/lib/user";
-import { AdminSidebar } from "./sidebar";
 
 type Props = {
   children: React.ReactNode;
@@ -15,6 +15,37 @@ type Props = {
 export const metadata = {
   title: "Admin",
 } satisfies Metadata;
+
+const adminRoutes = [
+  {
+    href: "/admin",
+    label: "Dashboard",
+  },
+  {
+    href: "/admin/tilbakemeldinger",
+    label: "Tilbakemeldinger",
+  },
+  {
+    href: "/admin/brukere",
+    label: "Brukere",
+  },
+  {
+    href: "/admin/happenings",
+    label: "Happenings",
+  },
+  {
+    href: "/admin/grupper",
+    label: "Grupper",
+  },
+  {
+    href: "/admin/studieretninger",
+    label: "Studieretninger",
+  },
+  {
+    href: "/admin/whitelist",
+    label: "Whitelist",
+  },
+];
 
 export default async function AdminDashboardLayout({ children }: Props) {
   const user = await getAuth();
@@ -27,7 +58,7 @@ export default async function AdminDashboardLayout({ children }: Props) {
     <div className="flex min-h-screen flex-col">
       <SiteHeader />
       <div className="flex w-full flex-grow flex-row">
-        <AdminSidebar>{children}</AdminSidebar>
+        <SidebarLayout routes={adminRoutes}>{children}</SidebarLayout>
       </div>
       <Footer />
     </div>
