+++ apps/web/src/app/(default)/auth/profil/arrangementer/page.tsx
@@ -1,0 +1,57 @@
+import Link from "next/link";
+import { redirect } from "next/navigation";
+
+import { getAuth } from "@echo-webkom/auth";
+import {
+  happeningTypeToPath,
+  happeningTypeToString,
+  registrationStatusToString,
+} from "@echo-webkom/lib";
+
+import { Chip } from "@/components/typography/chip";
+import { Heading } from "@/components/typography/heading";
+import { getUserRegistrations } from "@/lib/queries/user";
+
+export default async function UserHappenings() {
+  const user = await getAuth();
+
+  if (!user) {
+    return redirect("/auth/logg-inn");
+  }
+  const registrations = await getUserRegistrations(user.id);
+
+  return (
+    <div className="max-w-2xl">
+      <Heading level={2} className="mb-4">
+        Dine arrangementer
+      </Heading>
+      {registrations.length > 0 ? (
+        <ul className="flex flex-col divide-y">
+          {registrations.map((registration) => (
+            <li key={registration.happening.slug}>
+              <div className="py-3">
+                <Link
+                  href={
+                    happeningTypeToPath[registration.happening.type] +
+                    "/" +
+                    registration.happening.slug
+                  }
+                  className="text-lg font-semibold hover:underline"
+                >
+                  {registration.happening.title}
+                </Link>
+
+                <div className="mt-3 flex gap-1">
+                  <Chip>{happeningTypeToString[registration.happening.type]}</Chip>
+                  <Chip>{registrationStatusToString[registration.status]}</Chip>
+                </div>
+              </div>
+            </li>
+          ))}
+        </ul>
+      ) : (
+        <p>Du er ikke påmeldt noen arrangementer.</p>
+      )}
+    </div>
+  );
+}
